test-commit: upload test results as an artifact
This worked reasonably well until I started testing with pull requests
submitted from forked repositories.  GitHub doesn't provide great
granularity for gihub.token in this case:

- The base repo can't write statuses to the forked repo
- By default all write permissions will be stripped before a
  pull_request event triggered workflow runs.  So we can't just add
    permissions:
        statuses: write
- We can modify private repositories to 'Send write tokens to workflows
  from fork pull requests', but not anything more granular, which would
  make it easy to pass more then permission to write statuses by
  mistake.
- We create a Personal Access Token with only permission to read
  content and write statuses and pass that in instead of github.token.
  That's a bit clunky, but it works.

However, the intended way to accomplish something like this where we
want to run potentially malicious code from a forked repository and also
write to the upstream repository is apparently to use 'on:
workflow_run'.
https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/

So, following the pattern set there, change test-commit to upload
artifacts with the status of each test run on a commit.