expand-matrix: improved method for fetching every commit in a series
Improve upon `get-pr-sha` by instead taking a test matrix as input, and
expanding it with the full and short SHA of every commit in the series.
This enables more control of the matrix by enabling the caller to add
more metadata, like test name, that shouldn't result in another
matrix dimension.

Passing both the full and short SHA allows the full SHA to be passed as
`ref` to actions/checkout while the short SHA is used to label the test
and commit status context.