test-commit: attempt to set direct status url
In https://github.com/orgs/community/discussions/8945#discussioncomment
GitHub added the `job.check_run_id` which can be used to directly link
to the webpage for the currently running job.  However, it's gated
behind having read permissions on actions.  So we'll attempt to get the
html_url and fall back to the url for the current run.